+++ src/config/text/admin_texts.py
@@ -4,12 +4,12 @@ from datetime import datetime
 class AdminMessages:
     """These messages are used for admin chat."""
 
-    BOT_STARTUP_MESSAGE = {
+    BOT_STARTUP_ADMIN_MESSAGE = {
         "en": f"*Bot started* ({datetime.now().strftime('%d-%m-%Y %H:%M:%S')})\n",
         "uk": f"*Бота запущено* ({datetime.now().strftime('%d-%m-%Y %H:%M:%S')})\n"
     }
 
-    BOT_SHUTDOWN_MESSAGE = {
+    BOT_SHUTDOWN_ADMIN_MESSAGE = {
         "en": f"*Bot stopped* ({datetime.now().strftime('%d-%m-%Y %H:%M:%S')})\n",
         "uk": f"*Бота зупинено* ({datetime.now().strftime('%d-%m-%Y %H:%M:%S')})\n"
     }