+++ src/config/text/group_texts.py
@@ -169,3 +169,5 @@ class GroupMessages:
         "en": "*Looks like you are the only admin of this group chat.*",
         "uk": "*Ви єдиний адміністратор цієї групи.*"
     }
+
+