+++ src/middlewares/__init__.py
@@ -1,0 +1,1 @@
+from .startup_shutdown_middleware import StartupShutdownMiddleware