+++ CHANGELOG.md
@@ -3,13 +3,14 @@ All notable changes to this project will be documented here.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [Unreleased] - 2024-##-##
 ### Added
 - `CHANGELOG.md` and `CODE_OF_CONDUCT.md` created.
 - Ruff code linter integrated.
 - `ADMIN_CHAT_ID` added to `.env`.
 - `admin_router` for admin commands and "ping-pong" health check.
-- Startup/shutdown messages for user's groups and admin-reports.
+- Add new middleware: startup/shutdown messages for user's groups and admin-reports.
+- Group size cheacking. Restricting some commands in groups with 500+ users.
 ### Changed
 - License switched from MIT to GNU GPL v3.0.
 - Improved `pyproject.toml` and documentation.